+++ src/main/java/domain/piece/Cannon.java
@@ -1,5 +1,9 @@
-package domain;
+package domain.piece;
 
+import domain.Board;
+import domain.Direction;
+import domain.Position;
+import domain.Team;
 import java.util.HashSet;
 import java.util.Set;
