+++ src/main/java/domain/piece/Guard.java
@@ -1,28 +1,24 @@
-package domain;
+package domain.piece;
 
+import domain.Board;
+import domain.Direction;
+import domain.Position;
+import domain.Team;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-public class Solider extends Piece {
+public class Guard extends Piece {
 
-    public Solider(final Position position, final Team team, final Board board) {
+    public Guard(final Position position, final Team team, final Board board) {
         super(position, team, board);
     }
 
     @Override
     protected Set<Position> getMovablePositions() {
         return Direction.getStraightDirection().stream()
-                .filter(direction -> getUnmovableDirection() != direction)
                 .map(direction -> position.nextPosition(direction))
                 .filter(this::isMovable)
                 .collect(Collectors.toSet());
-    }
-
-    private Direction getUnmovableDirection() {
-        if (team == Team.BLUE) {
-            return Direction.BOTTOM;
-        }
-        return Direction.TOP;
     }
 
     private boolean isMovable(final Position position) {
