+++ src/test/java/domain/piece/PieceTest.java
@@ -1,9 +1,11 @@
-package domain;
+package domain.piece;
 
-import static org.assertj.core.api.Assertions.*;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
+import domain.Board;
+import domain.Position;
+import domain.Team;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
